requirements.txt
shioaji>=1.2.5
python-socketio[client]>=5.11.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
